Continue making little steps;
This patch adds the ability to call the Structurizr-CLI from within the
Rust code. It's not working as expected. It seems like the response from
the Docker Exec is getting cached? This might not be an issue if I
implement the ability for the Docker call to read dynamic input and if
the output is updated every time I run it. So we'll see. But for now, I
think I need to understand more what Bollard is doing or use some other
Crate for the Docker API.