Improving the error handling for --workspace;
This error handling is checking that the file that gets passed in as
workspace is a `.dsl` file and that it is an actual file that actually
exists. The error messaging needs to be improved for sure. I'm going to
be try to leverage the `anyhow` crate for this in the future.