Making things more legible for `run_export()`;
Previously, this was a `f` variable which wasn't very legible. I'm
updating this so that it follows a convention of `valid_*` before
passing these `&str` to the `run_export()` function. I like this
approach better as it allows me to allow for some testing in the main.rs
file to verify my own logic around this stuff. But that's a ways off at
this point. For now the work is to get this working and have variables
be legible to me later.